requirements.txt
dlib==19.24.0
Flask==2.2.2
keras==2.11.0
Keras_Preprocessing==1.1.2
numpy==1.21.6
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
pdf2image==1.16.0
Pillow==9.4.0
pillow-avif-plugin==1.3.1
tensorflow==2.11.0
tensorflow_intel==2.11.0
tqdm==4.64.1

